tests/wpt/web-platform-tests/preload/avoid-prefetching-on-text-plain.html

<!DOCTYPE html>
<title>Ensures content delivered with Content-Type: text/plain header is not prefetched</title>
<!-- Regression test for https://crbug.com/1160665 -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/preload/resources/preload_helper.js"></script>
<body>
    <script>
        setup({single_test: true});
        window.addEventListener("load", function() {
            verifyPreloadAndRTSupport();
            // This test works by loading a text/plain iframe containing a <script> tag.
            // It then injects some post-load JavaScript to serialize the Performance API
            // data and pass it back to this document.
            var prefetchingIframe = document.getElementById('prefetching-frame');
            window.addEventListener("message", function(msg) {
                // Parse the Performance API data passed from the plain text iframe.
                const entries = JSON.parse(msg.data);
                const urls = [];
                const resource_types = [];
                for (const entry of entries) {
                    resource_types.push(entry.entryType);
                    urls.push(entry.name);
                }
                // If preloading is working correctly, should only see the text document
                // represented in the performance information.
                assert_array_equals(resource_types, ['navigation']);
                assert_equals(urls.length, 1);
                assert_equals(urls[0].endsWith("avoid-prefetching-on-text-plain-inner.html"), true);
                done();
            });
            prefetchingIframe.addEventListener('load', function() {
                // Pass performance API info back to this document, process in above event handler.
                const passMsg = 'parent.postMessage(JSON.stringify(performance.getEntries()));';
                prefetchingIframe.contentWindow.eval(passMsg);
            });
            // Start the iframe load.
            prefetchingIframe.src = "avoid-prefetching-on-text-plain-inner.html";
        });
    </script>

    <iframe id="prefetching-frame"></iframe>
</body>
